+++ test/index.spec.js
@@ -1,0 +1,7 @@
+const index = require("../src/index");
+
+describe("index", () => {
+  it("should export Prismic", () => {
+    expect(index.Prismic).toBeDefined();
+  });
+});