+++ CHANGELOG.md
@@ -15,4 +15,4 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ### BREAKING CHANGES
 - Forked from xByOrange reactive-data-source v1.7.0 library. (Only origins.Prismic is exposed from now)
 - Extends from Mercury Origin instead of reactive-data-source origin. (Refer to mercury CHANLEGOG for further details)
-- Removed methods option. Now constructor accepts only two arguments.
+- Change constructor name from PrismicCMS to Prismic